Add regression test for bundle with size that is too large for int field. Change applications table migration to use bigint field for `bundle_size`

+++ migrations/201607110001_applications_up.sql
@@ -8,8 +8,8 @@ CREATE TABLE IF NOT EXISTS applications (
   identifier text,
   short_version text,
   version text,
-  bundle_size integer,
-  dynamic_size integer,
+  bundle_size bigint,
+  dynamic_size bigint,
   is_validated bool,
 
   install_count integer DEFAULT 1,
